Add initial HTML structure for Admin Dashboard pages including Help, Index, Settings, and Users

+++ signup/signup.js
@@ -5,6 +5,7 @@ const signUp = () => {
   if (firstName.value.trim() === '' || lastName.value.trim() === '' || email.value.trim() === '' || password.value.trim() === '' || termsCheck.value === false) {
     errorMessage.style.display = 'block'
     errorMessage2.style.display = 'none'
+    errorMessage3.style.display = 'none'
   } else {
     errorMessage.style.display = 'none'
     const userInfo = {
@@ -14,20 +15,35 @@ const signUp = () => {
       pass: password.value
     }
     let regexString = /^[a-zA-Z0-9._%+-]+@[a-zA-Z0-9.-]+\.[a-zA-Z]{2,}$/;
+    let PasswordString = /^(?=.*[a-z])(?=.*[A-Z])(?=.*\d)(?=.*[@$!%*?&])[A-Za-z\d@$!%*?&]{8,}$/;
     const confirmEmail = regexString.test(userInfo.mail);
+    const confirmPassword = PasswordString.test(userInfo.pass);
     if (confirmEmail) {
       const sameEmail = allUsers.find(user => user.mail === userInfo.mail);
       if (sameEmail) {
         alert('Account already exists')
       } else {
-        allUsers.push(userInfo);
-        localStorage.setItem('Tripco', JSON.stringify(allUsers));
-        console.log(allUsers);
-        window.location.href = '../signin/signin.html'
+        if (confirmPassword) {
+          allUsers.push(userInfo);
+          localStorage.setItem('Tripco', JSON.stringify(allUsers));
+          console.log(allUsers);
+          btn.innerHTML = `
+            <span class="spinner-grow spinner-grow-sm" aria-hidden="true"></span>
+            <span role="status">Loading ...</span>
+          `
+          setTimeout(() => {
+            window.location.href = '../signin/signin.html';
+          }, 2000)
+        } else {
+          errorMessage3.style.display = 'block'
+          errorMessage2.style.display = 'none';
+          errorMessage.style.display = 'none'
+        }
       }
     } else {
       errorMessage2.style.display = 'block';
       errorMessage.style.display = 'none'
+      errorMessage3.style.display = 'none'
     }
     firstName.value = ''
     lastName.value = ''
